restructure utils module
