+++ crates/katana-rpc/src/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod contract;
+pub mod transaction;
